refactor: keep service startup files within lint limits
